New approach removing all the validations and doing the mount before the start the container.
