+++ src/utils/SetupUtil.js
@@ -110,10 +110,6 @@ export default {
           machineVersion
         });
 
-        await virtualBox.getShareDir(machine.name());
-        await virtualBox.mountAllSharedDirs();
-
-
         let exists = await virtualBox.vmExists(machine.name()) && fs.existsSync(path.join(util.home(), '.docker', 'machine', 'machines', machine.name()));
         if (!exists) {
           router.get().transitionTo('setup');
